products loaded and domain configured

+++ src/models/product/ProductService.js
@@ -3,6 +3,10 @@ import { isValidBaimlProduct } from "@/utils/validate/validateBaimlProducts";
 import { isValidStoreProduct } from "@/utils/validate/validateStoreProduct";
 import { isValidProductInfo } from "@/utils/validate/validateProductCategoryInfo";
 import toNumericId from "@/utils/toNumericId";
+import {
+  deleteImageFromCloudinary,
+  uploadImagesToCloudinary,
+} from "@/utils/imageHandler/cloudinaryImagesHandler";
 
 class ProductService {
   constructor() {
@@ -47,10 +51,60 @@ class ProductService {
 
   async deleteProduct(productId) {
     try {
-      const deleteProduct = await this.dao.deleteProduct(
+      const product = await this.dao.findProductById(toNumericId(productId));
+      if (!product) throw new Error("El producto no existe");
+
+      //elimina todas las imagenes del producto de cloudinary solo cuando otro producto no este usando la misma imagen
+      //El false al final es para que no actualice la propiedad images del producto con la imagen eliminada , porque se eliminara el producto completo luego
+      for (const image of product.images) {
+        await this.deleteProductImage(productId, image.public_id, false);
+      }
+
+      const deletedProduct = await this.dao.deleteProduct(
         toNumericId(productId)
       );
-      return deleteProduct;
+
+      return deletedProduct;
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async checkIfProductImageIsShared(public_id) {
+    // Contar cuántos productos aún usan esta imagen
+    const count = await this.dao.countImages(public_id);
+
+    return count;
+  }
+
+  async deleteProductImage(productId, public_id, updateProduct) {
+    try {
+      const count = await this.checkIfProductImageIsShared(public_id);
+
+      if (count === 1) {
+        await deleteImageFromCloudinary(public_id);
+      }
+
+      if (updateProduct) {
+        return await this.dao.deleteProductImage(productId, public_id);
+      }
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async updateProductImage(productId, newImage) {
+    try {
+      const uploadResults = await uploadImagesToCloudinary([newImage]);
+      const uploadedImagesUrls = uploadResults.map((e) => ({
+        url: e.secure_url,
+        public_id: e.public_id,
+      }));
+
+      await this.dao.updateProductImage(
+        toNumericId(productId),
+        uploadedImagesUrls[0]
+      );
     } catch (error) {
       throw error;
     }
@@ -58,12 +112,11 @@ class ProductService {
 
   async updateProduct(productToUpdate, id) {
     try {
-      
       if (productToUpdate.kind === "Store") {
         isValidStoreProduct(productToUpdate);
       } else if (productToUpdate.kind === "Baiml") {
         isValidBaimlProduct(productToUpdate);
-      }else throw new Error("El tipo de producto no es valido");
+      } else throw new Error("El tipo de producto no es valido");
 
       const updateProduct = await this.dao.updateProduct(productToUpdate, id);
       return updateProduct;
